Fix Makefile, add some tests, dockerfile updated
