+++ Dockerfile
@@ -1,24 +1,24 @@
 FROM alpine:3.9.2
 
 RUN apk add --update --no-cache \ 
-    readline-dev libc-dev make gcc wget git zip unzip outils-md5
+    readline-dev libc-dev make gcc wget zip git unzip outils-md5 \
+    openssl openssl-dev
 
 ENV LUA_VERSION 5.3.5
 ENV LUAROCKS_VERSION 3.0.4
 
 RUN wget https://www.lua.org/ftp/lua-${LUA_VERSION}.tar.gz -O - | tar -xzf -
 WORKDIR /lua-$LUA_VERSION
-RUN make -j"$(nproc)" linux; make install
+RUN make linux test; make install
 WORKDIR /
 RUN rm -rf /lua-${LUA_VERSION}
 
 RUN wget https://luarocks.github.io/luarocks/releases/luarocks-${LUAROCKS_VERSION}.tar.gz -O - | tar -xzf -
 WORKDIR /luarocks-$LUAROCKS_VERSION
-RUN ./configure; make -j"$(nproc)" bootstrap
+RUN ./configure; make build; make install
 WORKDIR /
 RUN rm -rf /luarocks-$LUAROCKS_VERSION
 
-RUN apk add --update --no-cache openssl openssl-dev 
 RUN luarocks install luasocket 
 RUN luarocks install luaossl
 RUN luarocks install LuaSec
@@ -47,7 +47,7 @@ COPY lua/splay/ ./splay
 
 ## Launch Testing to verify the installation
 COPY lua/tests ./
-RUN busted ./ -p spec_
+RUN busted ./busted_specs
 
 ## Deploy script 
 COPY ./deploy.sh ./
