+++ c/data_bits.c
@@ -25,6 +25,7 @@
 #include <stdlib.h>
 #include <time.h>
 #include <sys/time.h>
+#include <sys/types.h>
 #include <stdio.h>
 #include <math.h>
 #include <string.h>
